arch: Added Player.cs and Enemy.cs classes
