+++ Assets/Prefabs/EnemyPrefab.prefab
@@ -75,9 +75,9 @@ GameObject:
   - component: {fileID: 9558302}
   - component: {fileID: -8217979379499544097}
   - component: {fileID: 8991707992796629231}
+  - component: {fileID: 2900200222520745750}
   - component: {fileID: -3239643052649976255}
   - component: {fileID: 8310605202814882490}
-  - component: {fileID: -7771885570771296760}
   - component: {fileID: 4390708264869013475}
   - component: {fileID: 8530034904484776134}
   m_Layer: 7
@@ -174,6 +174,22 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
+--- !u!114 &2900200222520745750
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 106386}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 949907ab6ee495a4e8c6299758f61cbb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  MovingController: {fileID: -3239643052649976255}
+  AnimationController: {fileID: 8310605202814882490}
+  HealthBarController: {fileID: 4390708264869013475}
+  KilledEnemiesUpdater: {fileID: 8530034904484776134}
 --- !u!114 &-3239643052649976255
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -187,9 +203,8 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _agent: {fileID: 5145800279972073281}
-  _chanceToBeTheOne: 0.5
+  _chanceToRunToTheExit: 0.5
   _enemyTarget: {x: 25, y: 0, z: 0}
-  _pointProvider: {fileID: -7771885570771296760}
 --- !u!114 &8310605202814882490
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -205,20 +220,6 @@ MonoBehaviour:
   _animator: {fileID: 9558302}
   _enemy: {fileID: -3239643052649976255}
   _healthBarController: {fileID: 4390708264869013475}
---- !u!114 &-7771885570771296760
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 106386}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fcea68804f8967f4d84e029621f02786, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  _terrain: {fileID: 15600000, guid: fb03f89b5cfec3c4593b737d6b4151da, type: 2}
-  _offsetFromTerrainEdges: 35
 --- !u!114 &4390708264869013475
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -234,8 +235,6 @@ MonoBehaviour:
   _gameLogic: {fileID: 0}
   _damageByHitWithPlayer: 30
   _labelUpdater: {fileID: 8530034904484776134}
-  MaxHealthPoints: 100
-  CurrentHealthPoints: 0
 --- !u!114 &8530034904484776134
 MonoBehaviour:
   m_ObjectHideFlags: 0
