+++ Assets/Scripts/Bonuses/AttackDamageBonus.cs
@@ -1,14 +1,18 @@
-﻿using UnityEngine;
+﻿using PlayerScripts;
+using UnityEngine;
 
-public class AttackDamageBonus : Bonus
+namespace Bonuses
 {
-    [SerializeField] 
-    private float _damageToAdd = 25f;
-    
-    public override void ApplyEffect(GameObject player)
+    public class AttackDamageBonus : Bonus
     {
-        var shootingController = player.GetComponent<PlayerShootingController>();
-        shootingController.minDamage += _damageToAdd;
-        shootingController.maxDamage += _damageToAdd;
+        [SerializeField] 
+        private float _damageToAdd = 25f;
+    
+        public override void ApplyEffect(Player player)
+        {
+            var shootingController = player.ShootingController;
+            shootingController.minDamage += _damageToAdd;
+            shootingController.maxDamage += _damageToAdd;
+        }
     }
 }
