+++ Assets/Scripts/Bonuses/Bonus.cs
@@ -1,9 +1,13 @@
+using PlayerScripts;
 using UnityEngine;
 
-public abstract class Bonus : MonoBehaviour
+namespace Bonuses
 {
-    public Sprite Image;
-    public string Description;
+    public abstract class Bonus : MonoBehaviour
+    {
+        public Sprite Image;
+        public string Description;
     
-    public abstract void ApplyEffect(GameObject target);
+        public abstract void ApplyEffect(Player target);
+    }
 }
