+++ Assets/Scripts/Bonuses/BonusView.cs
@@ -1,24 +1,28 @@
+using PlayerScripts;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class BonusView : MonoBehaviour
+namespace Bonuses
 {
-    private Bonus _bonus;
+    public class BonusView : MonoBehaviour
+    {
+        private Bonus _bonus;
 
-    [SerializeField]
-    private Image _image;
-    [SerializeField]
-    private Text _description;
+        [SerializeField]
+        private Image _image;
+        [SerializeField]
+        private Text _description;
     
-    public void Initialize(Bonus bonus)
-    {
-        _image.sprite = bonus.Image;
-        _description.text = bonus.Description;
-        _bonus = bonus;
-    }
+        public void Initialize(Bonus bonus)
+        {
+            _image.sprite = bonus.Image;
+            _description.text = bonus.Description;
+            _bonus = bonus;
+        }
 
-    public void UseBonus(GameObject target)
-    {
-        _bonus.ApplyEffect(target);
+        public void UseBonus(Player target)
+        {
+            _bonus.ApplyEffect(target);
+        }
     }
 }
