+++ Assets/Scripts/EnemyScripts/Enemy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 949907ab6ee495a4e8c6299758f61cbb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
