+++ Assets/Scripts/GameLogic.cs
@@ -1,9 +1,19 @@
+using PlayerScripts;
 using UI;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
 public class GameLogic : MonoBehaviour
 {
+    public static GameLogic Instance { get; private set; }
+
+    public Player Player;
+    
+    private void Awake()
+    {
+        Instance = this;
+    }
+
     public void Win()
     {
         Time.timeScale = 0;
