+++ Assets/Scripts/PlayerScripts/Player.cs
@@ -1,0 +1,15 @@
+using UI.HealthBar;
+using UnityEngine;
+
+namespace PlayerScripts
+{
+    public class Player : MonoBehaviour
+    {
+        public PlayerMovementController MovementController;
+        public PlayerAnimationController AnimationController;
+        public PlayerShootingController ShootingController;
+        public HealthBarController HealthBarController;
+
+        public Animator Animator;
+    }
+}
