+++ Assets/Scripts/Shard/ShardBounceManager.cs
@@ -1,27 +1,31 @@
+using Assets.Scripts.Utils;
 using UnityEngine;
 
-public class ShardBounceManager : MonoBehaviour
+namespace Shard
 {
-    [SerializeField] 
-    private float _maxBouncesCount = 3;
-    private float _bouncesCount;
+    public class ShardBounceManager : MonoBehaviour
+    {
+        [SerializeField] 
+        private float _maxBouncesCount = 3;
+        private float _bouncesCount;
     
-    private void Update()
-    {
-        if(_bouncesCount >= _maxBouncesCount)
-            Destroy(gameObject);
-    }
-
-    private void OnCollisionEnter(Collision other)
-    {
-        if (other.collider.CompareTag("BordersOfMap"))
+        private void Update()
         {
-            Destroy(gameObject);
-            return;
+            if(_bouncesCount >= _maxBouncesCount)
+                Destroy(gameObject);
         }
 
-        if (other.collider.CompareTag("Walls"))
-            _bouncesCount++;
-    }
+        private void OnCollisionEnter(Collision other)
+        {
+            if (other.collider.CompareTag(GlobalConstants.BordersOfMapTag))
+            {
+                Destroy(gameObject);
+                return;
+            }
 
+            if (other.collider.CompareTag(GlobalConstants.WallsTag))
+                _bouncesCount++;
+        }
+
+    }
 }
