+++ Assets/Scripts/Utils/TerrainPointProvider.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 using Random = UnityEngine.Random;
